fix: setup ios accept/decline events before user is connected
