+++ packages/react-native-sdk/src/hooks/push/useIosVoipPushEventsSetupEffect.ts
@@ -1,26 +1,14 @@
 import { type MutableRefObject, useEffect, useRef, useState } from 'react';
-import {
-  getCallKeepLib,
-  getVoipPushNotificationLib,
-} from '../../utils/push/libs';
+import { getVoipPushNotificationLib } from '../../utils/push/libs';
 
-import { AppState, Platform } from 'react-native';
-import { StreamVideoRN } from '../../utils';
+import { Platform } from 'react-native';
+import { onVoipNotificationReceived, StreamVideoRN } from '../../utils';
 import {
   useConnectedUser,
   useStreamVideoClient,
 } from '@stream-io/video-react-bindings';
 import { setPushLogoutCallback } from '../../utils/internal/pushLogoutCallback';
-import { NativeModules } from 'react-native';
-import {
-  canAddPushWSSubscriptionsRef,
-  shouldCallBeEnded,
-} from '../../utils/push/internal/utils';
-import {
-  pushUnsubscriptionCallbacks$,
-  voipPushNotificationCallCId$,
-} from '../../utils/push/internal/rxSubjects';
-import { RxUtils, StreamVideoClient, getLogger } from '@stream-io/video-client';
+import { StreamVideoClient, getLogger } from '@stream-io/video-client';
 
 const logger = getLogger(['useIosVoipPushEventsSetupEffect']);
 
@@ -59,6 +47,7 @@ export const useIosVoipPushEventsSetupEffect = () => {
   const lastVoipTokenRef = useRef({ token: '', userId: '' });
   const [unsentToken, setUnsentToken] = useState<string>();
 
+  // this effect is used to send the unsent token to stream
   useEffect(() => {
     const pushConfig = StreamVideoRN.getConfig().push;
     //  we need to wait for user to be connected before we can send the push token
@@ -96,9 +85,11 @@ export const useIosVoipPushEventsSetupEffect = () => {
 
   useEffect(() => {
     const pushConfig = StreamVideoRN.getConfig().push;
-    if (Platform.OS !== 'ios' || !pushConfig || !client) {
+    const pushProviderName = pushConfig?.ios.pushProviderName;
+    if (Platform.OS !== 'ios' || !pushConfig || !client || !pushProviderName) {
       return;
     }
+
     const voipPushNotification = getVoipPushNotificationLib();
 
     // even though we do this natively, we have to still register here again
@@ -124,16 +115,12 @@ export const useIosVoipPushEventsSetupEffect = () => {
         );
         return;
       }
-      const push_provider_name = pushConfig.ios.pushProviderName;
-      if (!push_provider_name) {
-        return;
-      }
       logger(
         'debug',
         `Sending voip token to stream, token: ${token} userId: ${userId}`
       );
       client
-        .addVoipDevice(token, 'apn', push_provider_name)
+        .addVoipDevice(token, 'apn', pushProviderName)
         .then(() => {
           logger(
             'debug',
@@ -152,10 +139,6 @@ export const useIosVoipPushEventsSetupEffect = () => {
       onTokenReceived(token);
     });
 
-    voipPushNotification.addEventListener('notification', (notification) => {
-      onNotificationReceived(notification);
-    });
-
     // this will fire when there are events occured before js bridge initialized
     voipPushNotification.addEventListener('didLoadWithEvents', (events) => {
       if (!events || !Array.isArray(events) || events.length < 1) {
@@ -166,7 +149,7 @@ export const useIosVoipPushEventsSetupEffect = () => {
         if (name === 'RNVoipPushRemoteNotificationsRegisteredEvent') {
           onTokenReceived(data);
         } else if (name === 'RNVoipPushRemoteNotificationReceivedEvent') {
-          onNotificationReceived(data);
+          onVoipNotificationReceived(data);
         }
       }
     });
@@ -185,124 +168,6 @@ export const useIosVoipPushEventsSetupEffect = () => {
       logger('debug', 'Voip event listeners are removed for user: ' + userId);
       voipPushNotification.removeEventListener('didLoadWithEvents');
       voipPushNotification.removeEventListener('register');
-      voipPushNotification.removeEventListener('notification');
     };
   }, [client]);
-};
-
-const onNotificationReceived = async (notification: any) => {
-  /* --- Example payload ---
-  {
-    "aps": {
-      "alert": {
-        "body": "",
-        "title": "Vishal Narkhede is calling you"
-      },
-      "badge": 0,
-      "category": "stream.video",
-      "mutable-content": 1
-    },
-    "stream": {
-      "call_cid": "default:ixbm7y0k74pbjnq",
-      "call_display_name": "",
-      "created_by_display_name": "Vishal Narkhede",
-      "created_by_id": "vishalexpo",
-      "receiver_id": "santhoshexpo",
-      "sender": "stream.video",
-      "type": "call.ring",
-      "version": "v2"
-    }
-  } */
-  const sender = notification?.stream?.sender;
-  const type = notification?.stream?.type;
-  // do not process any other notifications other than stream.video or ringing
-  if (sender !== 'stream.video' && type !== 'call.ring') {
-    return;
-  }
-  const call_cid = notification?.stream?.call_cid;
-  const pushConfig = StreamVideoRN.getConfig().push;
-  if (!call_cid || Platform.OS !== 'ios' || !pushConfig) {
-    return;
-  }
-  const client = await pushConfig.createStreamVideoClient();
-  if (!client) {
-    return;
-  }
-  const callFromPush = await client.onRingingCall(call_cid);
-  let uuid = '';
-  try {
-    uuid =
-      await NativeModules?.StreamVideoReactNative?.getIncomingCallUUid(
-        call_cid
-      );
-  } catch (error) {
-    logger('error', 'Error in getting call uuid from native module', error);
-  }
-  if (!uuid) {
-    logger(
-      'error',
-      `Not processing call.ring push notification, as no uuid found for call_cid: ${call_cid}`
-    );
-    return;
-  }
-  const created_by_id = notification?.stream?.created_by_id;
-  const receiver_id = notification?.stream?.receiver_id;
-  function closeCallIfNecessary() {
-    const { mustEndCall, callkeepReason } = shouldCallBeEnded(
-      callFromPush,
-      created_by_id,
-      receiver_id
-    );
-    if (mustEndCall) {
-      const callkeep = getCallKeepLib();
-      logger(
-        'debug',
-        `callkeep.reportEndCallWithUUID for uuid: ${uuid}, call_cid: ${call_cid}, reason: ${callkeepReason}`
-      );
-      callkeep.reportEndCallWithUUID(uuid, callkeepReason);
-      const voipPushNotification = getVoipPushNotificationLib();
-      voipPushNotification.onVoipNotificationCompleted(uuid);
-      return true;
-    }
-    return false;
-  }
-  const closed = closeCallIfNecessary();
-  const canListenToWS = () =>
-    canAddPushWSSubscriptionsRef.current && AppState.currentState !== 'active';
-  if (!closed && canListenToWS()) {
-    const unsubscribe = callFromPush.on('all', (event) => {
-      const _canListenToWS = canListenToWS();
-      if (!_canListenToWS) {
-        logger(
-          'debug',
-          `unsubscribe due to event callCid: ${call_cid} canListenToWS: ${_canListenToWS}`,
-          event
-        );
-        unsubscribe();
-        return;
-      }
-      const _closed = closeCallIfNecessary();
-      if (_closed) {
-        logger(
-          'debug',
-          `unsubscribe due to event callCid: ${call_cid} canListenToWS: ${_canListenToWS} shouldCallBeClosed: ${_closed}`,
-          event
-        );
-        unsubscribe();
-      }
-    });
-    const unsubscriptionCallbacks =
-      RxUtils.getCurrentValue(pushUnsubscriptionCallbacks$) ?? [];
-    pushUnsubscriptionCallbacks$.next([
-      ...unsubscriptionCallbacks,
-      unsubscribe,
-    ]);
-  }
-  // send the info to this subject, it is listened by callkeep events
-  // callkeep events will then accept/reject the call
-  logger(
-    'debug',
-    `call_cid:${call_cid} uuid:${uuid} received and processed from call.ring push notification`
-  );
-  voipPushNotificationCallCId$.next(call_cid);
 };
